+++ webargs/fields.py
@@ -3,6 +3,16 @@
 
 Includes all fields from `marshmallow.fields` in addition to a custom
 `Nested` field.
+
+All fields can optionally take a special `location` keyword argument, which tells webargs
+where to parse the request argument from. ::
+
+    args = {
+        'active': fields.Bool(location='query')
+        'content_type': fields.Str(load_from='Content-Type',
+                                   location='headers')
+    }
+
 """
 import marshmallow as ma